Move switch under org commands * Fix arg check for team ls cmd
