+++ cmd/amp/commands.go
@@ -9,7 +9,6 @@ import (
 	"github.com/appcelerator/amp/cli/command/org"
 	"github.com/appcelerator/amp/cli/command/password"
 	"github.com/appcelerator/amp/cli/command/stats"
-	"github.com/appcelerator/amp/cli/command/switch"
 	"github.com/appcelerator/amp/cli/command/team"
 	"github.com/appcelerator/amp/cli/command/user"
 	"github.com/appcelerator/amp/cli/command/version"
@@ -78,9 +77,6 @@ func addCommands(cmd *cobra.Command, c cli.Interface) {
 
 		//password
 		password.NewPasswordCommand(c),
-
-		//switch
-		switch_.NewSwitchCommand(c),
 
 		//team
 		team.NewTeamCommand(c),
